move client & logic into api module
